% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hyperlink.R
\name{get_links}
\alias{get_links}
\title{List if URLs to create hyperlinks}
\usage{
get_links(customer = "CHI", project.id = 12396)
}
\arguments{
\item{customer}{The customer id from secuTrial}

\item{project.id}{The project id from secuTrial obtained from the Admin Tool}
}
\description{
This function contains links to ..
}
