% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sae.R
\docType{data}
\name{sae_severity_list}
\alias{sae_severity_list}
\alias{mod_sae_server}
\title{Serious Adverse Events}
\format{
An object of class \code{character} of length 3.
}
\usage{
sae_severity_list

mod_sae_server(input, output, session, data.sae)
}
\arguments{
\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data.sae}{data for use in calculations}

\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}
}
\description{
In development
}
\author{
Laurent Brodier
}
\keyword{datasets}
